Issue #2031, Moved logic, moved send btn texts to another Class
