+++ extension/chrome/elements/compose.ts
@@ -392,7 +392,7 @@ Catch.try(async () => {
   BrowserMsg.addListener('passphrase_entry', async ({ entered }: Bm.PassphraseEntry) => {
     if (!entered) {
       clearInterval(passphraseInterval);
-      composer.resetSendBtn();
+      composer.composerSendBtn.resetSendBtn();
     }
   });
   BrowserMsg.listen(tabId);
